FName: We now support constructing with FNAME_Add instead of FNAME_Find.
Nothing had to be done to support this other than adding the
parameter to the constructors.

The default is still FNAME_Find, even though the default is
FNAME_Add in the UE source.
This should maybe change so that it matches the UE source,
I'm too lazy to make sure that it works exactly like I want
it to with that change.

Note that this could be tricky to support if we replace
the FName constructor AOB with the names array.